create/delete vacancy with userId

+++ src/modules/Vacancy/vacancy.service.js
@@ -1,9 +1,9 @@
 import Vacancy from "./vacancy.model.js";
 
-export async function createVacancyService(body) {
-  const newVacany = new Vacancy(body);
-  await newVacany.save();
-  return newVacany;
+export async function createVacancyService(body, userId) {
+  const vacancy = new Vacancy({ ...body, recruiter: userId });
+  await vacancy.save();
+  return vacancy;
 }
 
 export async function getVacanciesService() {
@@ -19,8 +19,16 @@ export async function getVacancyByIdService(vacancyId) {
   return vacancy;
 }
 
-export async function deleteVacancyService(vacancyId) {
-  const vacancy = await getVacancyByIdService(vacancyId);
+export async function deleteVacancyService(vacancyId, userId) {
+  const vacancy = await Vacancy.findOne({
+    _id: { $eq: vacancyId },
+    recruiter: { $eq: userId },
+  });
+
+  if (!vacancy) {
+    throw new Error("Vacancy not found");
+  }
+
   await vacancy.deleteOne();
   return vacancy;
 }
